Added Further Reading (#1417)
Hey there - I was looking into Polygon's PoS and lacked a technical understanding of the architecture and how to deploy a node. I found the above links extremely useful and adding them to the page will ease navigation and usability for developers. They strengthen and build upon the content provided in the rest of the page.